re-arrange some things and also make a more compact gui version
Consolidated things into the safari_detect.py, which is a gui style inspired by https://github.com/rhmatti/CUEBIT-Spectrum-Analyzer.

Made adjustments to the previous detect processing scripts so that they can work for both this new gui as well as the old one. This mostly involved having flags for disabling the fig.show().