Adjustments to allow fitting shoulders as well.
This uses 2 applications of numpy.gradient to produce a second derivative for the data to fit. This allows for locating the shoulders in the plots. This derivative also has the gaussian binning algorithm applied to it to smooth out some of the numerical errors that occur durng this derivative process.